+++ uefi-test-runner/src/main.rs
@@ -1,6 +1,7 @@
 #![no_std]
 #![no_main]
 #![feature(abi_efiapi)]
+#![feature(negative_impls)]
 
 #[macro_use]
 extern crate log;